match fix

+++ commands/match.js
@@ -5,16 +5,30 @@ module.exports = async (message) => {
   ${matchcreator} Is looking for a match! 
   React with ⚔ to notify them for a match!
   ` 
-  let messageSend = await message.channel.send(text)
+
+  if(message.guild.id != 579651261054451752) return message.channel.send('This command is only avaible in Helpmet Competitive!')
+  
+  let messageSend = await message.guild.channels.find(c => c.id == 586242295091953684).send(text)
+  
 
   messageSend.react('⚔')
 
   const filter = (reaction, user) => reaction.emoji.name === '⚔' && !user.bot
-  const collector = messageSend.createReactionCollector(filter, { max: 1 })
-  collector.on('collect', (reaction, reason) => {
+  const collector = messageSend.createReactionCollector(filter, { max: 1, time: 3600*1000 })
+  collector.on('collect', async (reaction, reason) => {
     let user = reaction.users.last()
-    message.channel.send(`**${user.username+'#'+user.discriminator}** Notified ${matchcreator} that a match is awaiting them!`)
     messageSend.delete()
-    message.author.send(`Looks like you found a match in <#${message.channel.id}> . Against **${user.username+'#'+user.discriminator}**`)
+    if(user.id == message.author.id) {
+      let canceled = await messageSend.channel.send(`${matchcreator} Canceled match searching!`)
+      setTimeout(() => { canceled.delete() }, 5000)
+      return
+    }
+    let notified = await messageSend.channel.send(`**${user.username+'#'+user.discriminator}** Notified ${matchcreator} that a match is awaiting them!`)
+    message.author.send(`Looks like you found a match in **Helmet Competitive** . Against **${user.username+'#'+user.discriminator}**`)
+
+    setTimeout(() => { notified.delete() }, 5000)
+  })
+  collector.on('end', async (reaction, reason) => {
+    messageSend.delete()
   })
 }
